taint/test/mochitest/test_object.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test Object Sink</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let string_content = "hello";
      let sink_names = [
          "object.data"
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          SimpleTest.is(report.detail.str, "tainted=hello", "Check sink string content");

          let flow = report.detail.str.taint[0].flow;
          SimpleTest.is(flow[2].operation, sink_names[i]);

          i += 1;
          if (i >= sink_names.length) {
              SimpleTest.finish();
          }
      }, false);

      function startTest() {
          let obj = document.getElementById("obj");
          let tainted = String.tainted("tainted=hello");

          obj.data = tainted;
      }
      
    </script>
  </head>

  <body onload="startTest();">
      <object id="obj"></object>
  </body>

</html>
